pyproject.toml
[project]
name = "test-backend-1-weit145"
version = "0.1.0"
description = ""
authors = [
    {name = "Weit145",email = "weit.dima@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "ruff (>=0.15.7,<0.16.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "freezegun (>=1.5.5,<2.0.0)"
]

[tool.pytest.ini_options]
pythonpath = "."
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-cov (>=7.1.0,<8.0.0)"
]
